+++ spp_drims/wizard/__init__.py
@@ -9,3 +9,4 @@ from . import allocation_preview_wizard
 from . import request_from_template_wizard
 from . import create_return_wizard
 from . import inspection_wizard
+from . import receive_wizard